certificates/pyproject.toml
[tool.poetry]
name = "certificates"
version = "0.1.0"
description = ""
authors = ["matias-gonz <maigonzalez@fi.uba.ar>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
py-solc-x = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
